Added `maintenance_device` method to devices.py
Added `maintenance_device` method to correspond with this API call:

https://docs.librenms.org/API/Devices/#maintenance_device

Tested and functions as expected. 

Kept duration as a string to keep the "H:m" formatting consistent. Initially thought about making duration an integer in minutes, then did math to convert, but keeping string seems to be the best approach here since the API appears to be looking for a string.